Implemented CustomFoodSearchActivity to pull data from Firebase and show on recyclerView
